+++ blocks/order-details/order-details.css
@@ -48,6 +48,10 @@
   height: 60px !important
 }
 
+.order-details-wrapper :is(.h-10) {
+  height: 2.5rem
+}
+
 .order-details-wrapper :is(.h-\[100px\]) {
   height: 100px
 }
@@ -68,8 +72,40 @@
   width: 60px !important
 }
 
+.order-details-wrapper :is(.w-16) {
+  width: 4rem
+}
+
+.order-details-wrapper :is(.w-20) {
+  width: 5rem
+}
+
+.order-details-wrapper :is(.w-28) {
+  width: 7rem
+}
+
+.order-details-wrapper :is(.w-36) {
+  width: 9rem
+}
+
+.order-details-wrapper :is(.w-48) {
+  width: 12rem
+}
+
+.order-details-wrapper :is(.w-52) {
+  width: 13rem
+}
+
+.order-details-wrapper :is(.w-80) {
+  width: 20rem
+}
+
 .order-details-wrapper :is(.w-\[100px\]) {
   width: 100px
+}
+
+.order-details-wrapper :is(.w-\[150px\]) {
+  width: 150px
 }
 
 .order-details-wrapper :is(.w-\[24px\]) {
@@ -88,40 +124,8 @@
   width: 100%
 }
 
-.order-details-wrapper :is(.w-20) {
-  width: 5rem
-}
-
-.order-details-wrapper :is(.w-36) {
-  width: 9rem
-}
-
-.order-details-wrapper :is(.w-80) {
-  width: 20rem
-}
-
-.order-details-wrapper :is(.w-16) {
-  width: 4rem
-}
-
-.order-details-wrapper :is(.w-28) {
-  width: 7rem
-}
-
-.order-details-wrapper :is(.w-48) {
-  width: 12rem
-}
-
-.order-details-wrapper :is(.w-52) {
-  width: 13rem
-}
-
-.order-details-wrapper :is(.w-9) {
-  width: 2.25rem
-}
-
-.order-details-wrapper :is(.w-\[150px\]) {
-  width: 150px
+.order-details-wrapper :is(.w-\[325px\]) {
+  width: 325px
 }
 
 .order-details-wrapper :is(.flex-1) {
@@ -168,12 +172,20 @@
   gap: 0.5rem
 }
 
+.order-details-wrapper :is(.gap-2\.5) {
+  gap: 0.625rem
+}
+
 .order-details-wrapper :is(.gap-28) {
   gap: 7rem
 }
 
 .order-details-wrapper :is(.gap-3) {
   gap: 0.75rem
+}
+
+.order-details-wrapper :is(.gap-3\.5) {
+  gap: 0.875rem
 }
 
 .order-details-wrapper :is(.gap-4) {
@@ -188,14 +200,6 @@
   gap: 1.5rem
 }
 
-.order-details-wrapper :is(.gap-2\.5) {
-  gap: 0.625rem
-}
-
-.order-details-wrapper :is(.gap-3\.5) {
-  gap: 0.875rem
-}
-
 .order-details-wrapper :is(.self-stretch) {
   align-self: stretch
 }
@@ -208,12 +212,12 @@
   border-radius: 9999px
 }
 
-.order-details-wrapper :is(.border-2) {
-  border-width: 2px
-}
-
 .order-details-wrapper :is(.border) {
   border-width: 1px
+}
+
+.order-details-wrapper :is(.border-2) {
+  border-width: 2px
 }
 
 .order-details-wrapper :is(.border-b) {
@@ -275,6 +279,11 @@
   background-color: rgb(243 244 246 / var(--tw-bg-opacity))
 }
 
+.order-details-wrapper :is(.bg-gray-50) {
+  --tw-bg-opacity: 1;
+  background-color: rgb(249 250 251 / var(--tw-bg-opacity))
+}
+
 .order-details-wrapper :is(.bg-green-50) {
   --tw-bg-opacity: 1;
   background-color: rgb(240 253 244 / var(--tw-bg-opacity))
@@ -290,17 +299,20 @@
   background-color: rgb(255 255 255 / var(--tw-bg-opacity))
 }
 
-.order-details-wrapper :is(.bg-gray-50) {
-  --tw-bg-opacity: 1;
-  background-color: rgb(249 250 251 / var(--tw-bg-opacity))
-}
-
 .order-details-wrapper :is(.fill-current) {
   fill: currentColor
 }
 
-.order-details-wrapper :is(.p-5) {
-  padding: 1.25rem
+.order-details-wrapper :is(.p-2) {
+  padding: 0.5rem
+}
+
+.order-details-wrapper :is(.p-2\.5) {
+  padding: 0.625rem
+}
+
+.order-details-wrapper :is(.p-3) {
+  padding: 0.75rem
 }
 
 .order-details-wrapper :is(.p-6) {
@@ -311,16 +323,8 @@
   padding: 18px
 }
 
-.order-details-wrapper :is(.p-3) {
-  padding: 0.75rem
-}
-
-.order-details-wrapper :is(.p-2) {
-  padding: 0.5rem
-}
-
-.order-details-wrapper :is(.p-2\.5) {
-  padding: 0.625rem
+.order-details-wrapper :is(.p-\[20px\]) {
+  padding: 20px
 }
 
 .order-details-wrapper :is(.px-2) {
@@ -333,6 +337,11 @@
   padding-right: 0.625rem
 }
 
+.order-details-wrapper :is(.px-5) {
+  padding-left: 1.25rem;
+  padding-right: 1.25rem
+}
+
 .order-details-wrapper :is(.px-6) {
   padding-left: 1.5rem;
   padding-right: 1.5rem
@@ -341,11 +350,6 @@
 .order-details-wrapper :is(.py-2) {
   padding-top: 0.5rem;
   padding-bottom: 0.5rem
-}
-
-.order-details-wrapper :is(.px-5) {
-  padding-left: 1.25rem;
-  padding-right: 1.25rem
 }
 
 .order-details-wrapper :is(.py-3) {
@@ -367,10 +371,6 @@
 
 .order-details-wrapper :is(.font-\[\'Inter\'\]) {
   font-family: 'Inter'
-}
-
-.order-details-wrapper :is(.font-\[\'TWK_Lausanne_Pan\'\]) {
-  font-family: 'TWK Lausanne Pan'
 }
 
 .order-details-wrapper :is(.\!text-4xl) {
@@ -447,6 +447,10 @@
   line-height: 1
 }
 
+.order-details-wrapper :is(.leading-relaxed) {
+  line-height: 1.625
+}
+
 .order-details-wrapper :is(.leading-snug) {
   line-height: 1.375
 }
@@ -465,6 +469,11 @@
   color: rgb(117 35 255 / var(--tw-text-opacity))
 }
 
+.order-details-wrapper :is(.text-gray-500) {
+  --tw-text-opacity: 1;
+  color: rgb(107 114 128 / var(--tw-text-opacity))
+}
+
 .order-details-wrapper :is(.text-green-700) {
   --tw-text-opacity: 1;
   color: rgb(21 128 61 / var(--tw-text-opacity))
@@ -480,29 +489,8 @@
   color: rgb(255 255 255 / var(--tw-text-opacity))
 }
 
-.order-details-wrapper :is(.text-gray-500) {
-  --tw-text-opacity: 1;
-  color: rgb(107 114 128 / var(--tw-text-opacity))
-}
-
 .order-details-wrapper :is(.line-through) {
   text-decoration-line: line-through
-}
-
-.order-details-wrapper :is(.outline) {
-  outline-style: solid
-}
-
-.order-details-wrapper :is(.outline-1) {
-  outline-width: 1px
-}
-
-.order-details-wrapper :is(.outline-offset-\[-1px\]) {
-  outline-offset: -1px
-}
-
-.order-details-wrapper :is(.outline-gray-300) {
-  outline-color: #d1d5db
 }
 
 .order-details-wrapper :is(.filter) {
@@ -510,14 +498,6 @@
 }
 
 @media (min-width: 768px) {
-  .order-details-wrapper :is(.md\:w-\[20\%\]) {
-    width: 20%
-  }
-
-  .order-details-wrapper :is(.md\:w-\[30\%\]) {
-    width: 30%
-  }
-
   .order-details-wrapper :is(.md\:w-\[25\%\]) {
     width: 25%
   }
