+++ blocks/order-status/order-status.css
@@ -144,32 +144,16 @@
   width: 200px
 }
 
+.order-status-wrapper :is(.w-\[80\%\]) {
+  width: 80%
+}
+
 .order-status-wrapper :is(.w-\[930px\]) {
   width: 930px
 }
 
 .order-status-wrapper :is(.w-full) {
   width: 100%
-}
-
-.order-status-wrapper :is(.w-\[70\%\]) {
-  width: 70%
-}
-
-.order-status-wrapper :is(.w-\[75\%\]) {
-  width: 75%
-}
-
-.order-status-wrapper :is(.w-\[85\%\]) {
-  width: 85%
-}
-
-.order-status-wrapper :is(.w-\[05\%\]) {
-  width: 05%
-}
-
-.order-status-wrapper :is(.w-\[80\%\]) {
-  width: 80%
 }
 
 .order-status-wrapper :is(.min-w-\[415px\]) {
@@ -621,14 +605,6 @@
 }
 
 @media (min-width: 768px) {
-  .order-status-wrapper :is(.md\:w-\[20\%\]) {
-    width: 20%
-  }
-
-  .order-status-wrapper :is(.md\:w-\[30\%\]) {
-    width: 30%
-  }
-
   .order-status-wrapper :is(.md\:w-\[25\%\]) {
     width: 25%
   }
